dirs refactored
